+++ backend/src/message/models/message.model.ts
@@ -1,4 +1,4 @@
-import { Column, DataType, ForeignKey, Model } from "sequelize-typescript";
+import { Column, DataType, ForeignKey, Model, Table } from "sequelize-typescript";
 import { ApiProperty } from "@nestjs/swagger";
 import { ConversationModel } from "./conversation.model";
 import { UserModel } from "../../user/models/user.model";
@@ -10,6 +10,7 @@ interface IMessageModel {
   receiverId: number;
 }
 
+@Table({ tableName: "messages", createdAt: false, updatedAt: false })
 export class MessageModel extends Model<MessageModel, IMessageModel> {
   @ApiProperty({ example: "4", description: "Message's id" })
   @Column({ type: DataType.INTEGER, unique: true, autoIncrement: true, primaryKey: true })
@@ -17,7 +18,7 @@ export class MessageModel extends Model<MessageModel, IMessageModel> {
 
   @ApiProperty({ example: "Hello!", description: "Message's text" })
   @Column({ type: DataType.STRING })
-  text: number;
+  text: string;
 
   @ApiProperty({ example: "false", description: "Message's status" })
   @Column({ type: DataType.BOOLEAN, defaultValue: false })
@@ -32,9 +33,4 @@ export class MessageModel extends Model<MessageModel, IMessageModel> {
   @Column({ type: DataType.INTEGER })
   @ForeignKey(() => UserModel)
   senderId: number;
-
-  @ApiProperty({ example: "93", description: "Receiver id" })
-  @Column({ type: DataType.INTEGER })
-  @ForeignKey(() => UserModel)
-  receiverId: number;
 }